Attempted to switch to using a stored buffer. Oddly the draw order now looks completely different. So it’s possible there’s a lurking problem.

+++ src/OpenGL/DrawElementsBuffer.h
@@ -17,7 +17,7 @@ class DrawElementsBuffer
 		DrawElementsBuffer(GLenum indexType);
 		virtual ~DrawElementsBuffer();
 
-		int addIndex(void *index);
+		size_t addIndex(void *index);
 
 		void bind();
